refactor: クラスの抽出
