+++ src/api/presentation/TodoController.ts
@@ -1,4 +1,3 @@
-import express from "express";
 import { Body, Delete, Get, Patch, Path, Post, Put, Route, Tags } from "tsoa";
 import { TodoService } from "../application/TodoService";
 
@@ -28,7 +27,7 @@ export interface TodoRequest {
 const service = new TodoService();
 @Route("api")
 @Tags("todo")
-class TodoController {
+export class TodoController {
   private service: TodoService;
 
   constructor() {
@@ -55,8 +54,7 @@ class TodoController {
   }
 
   @Delete("/todo")
-  public async delete(@Body() req: any): Promise<void> {
-    const request = JSON.parse(JSON.stringify(req.body));
+  public async delete(@Body() request: any): Promise<void> {
     if (request.id !== null) await this.service.delete(request.id);
   }
 
@@ -65,65 +63,3 @@ class TodoController {
     return this.service.update(todo);
   }
 }
-const router = express.Router();
-const controller = new TodoController();
-
-router.get("/todos", async (req, res) => {
-  try {
-    const result = await controller.selectAll();
-    res.send(result);
-  } catch (err: any) {
-    res.status(400).send({ error: err.message });
-  }
-});
-
-router.get("/todos/count", async (req, res) => {
-  try {
-    const result = await controller.count();
-    res.send(result.toString());
-  } catch (err: any) {
-    res.status(400).send({ error: err.message });
-  }
-});
-
-router.get("/todo/:id", async (req, res) => {
-  try {
-    const id = parseInt(req.params.id);
-    const result = await controller.find(id);
-    res.send({ value: result });
-  } catch (err: any) {
-    res.status(400).send({ error: err.message });
-  }
-});
-
-router.post("/todo", async (req, res) => {
-  try {
-    const request: TodoRequest = req.body;
-    await controller.create(request);
-    res.send({ success: "success" });
-  } catch (err: any) {
-    res.status(400).send({ error: err.message });
-  }
-});
-
-router.delete("/todo", async (req, res) => {
-  try {
-    const request = JSON.parse(JSON.stringify(req.body));
-    await controller.delete(request);
-    res.send({ success: "success" });
-  } catch (err: any) {
-    res.status(400).send({ error: err.message });
-  }
-});
-
-router.put("/todo", async (req, res) => {
-  try {
-    const request: TodoRequest = req.body;
-    await controller.update(request);
-    res.send({ success: "success" });
-  } catch (err: any) {
-    res.status(400).send({ error: err.message });
-  }
-});
-
-export default router;
